Simulated data set from an illness-death model.

This function creates a single simulated data set for a single treatment arm. It simulates data from an illness-death model with one row per transition and subject.

getSimulatedData(
  N,
  transition = exponential_transition(h01 = 1, h02 = 1, h12 = 1),
  dropout = list(rate = 0, time = 12),
  accrual = list(param = "time", value = 0)
)

Arguments

N

(int)
number of patients.

transition

(TransitionParameters)
transition parameters comprising hazards, corresponding intervals and weibull_rates, see exponential_transition(), piecewise_exponential() and weibull_transition() for details.

dropout

(list)
specifies drop-out probability. Random censoring times are generated using exponential distribution. dropout$rate specifies the drop-out probability per dropout$time time units. If dropout$rate is equal to 0, then no censoring is applied.

accrual

(list)
specifies accrual intensity. See addStaggeredEntry() for details.

Value

This returns a data frame with one row per transition per individual.

Details

The output data set contains the following columns:

id (integer): patient id.

from (numeric): starting state of the transition.

to (character): final state of the transition.

entry (numeric): entry time of the transition on the individual time scale.

exit (numeric): exit time of the transition on the individual time scale.

entryAct (numeric): entry time of the transition on study time scale.

exitAct (numeric): exit time of the transition on study time scale.

censAct (numeric): censoring time of the individual on study time scale.

Examples

getSimulatedData(10,
  transition = exponential_transition(h01 = 1, h02 = 1.5, h12 = 1),
  dropout = list(rate = 0.3, time = 1),
  accrual = list(param = "time", value = 5)
)
#>    id from   to      entry       exit  entryAct   exitAct   censAct
#> 1   1    0    2 0.00000000 0.33219889 2.3409790 2.6731779 9.5885301
#> 2   2    0    1 0.00000000 0.01307901 4.9658175 4.9788965 8.6098678
#> 3   2    1    2 0.01307901 0.56922971 4.9788965 5.5350472 8.6098678
#> 4   3    0    2 0.00000000 0.28163880 3.9475461 4.2291849 8.5604559
#> 5   4    0    2 0.00000000 0.88358163 2.5607903 3.4443719 6.4969933
#> 6   5    0    2 0.00000000 0.16979297 0.5236883 0.6934813 0.7054999
#> 7   6    0    2 0.00000000 0.60767329 1.1663982 1.7740715 2.6714160
#> 8   7    0    2 0.00000000 0.75049952 2.1875497 2.9380492 4.1515061
#> 9   8    0 cens 0.00000000 0.29298878 2.7219952 3.0149840 3.0149840
#> 10  9    0    1 0.00000000 0.08403398 1.9624626 2.0464966 2.7376964
#> 11  9    1 cens 0.08403398 0.77523372 2.0464966 2.7376964 2.7376964
#> 12 10    0    1 0.00000000 0.49931941 2.4395420 2.9388614 4.8414456
#> 13 10    1 cens 0.49931941 2.40190358 2.9388614 4.8414456 4.8414456